college_bbs/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
Django==3.2.9
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
mysqlclient==2.1.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
